requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cached-property==2.0.1
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dash==3.2.0
dash-bootstrap-components==2.0.4
Flask==3.1.2
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mypy_extensions==1.1.0
narwhals==2.9.0
nest-asyncio==1.6.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
sqlalchemy-orm==1.2.10
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.23.0
